Refactor style component
