+++ modules/index.js
@@ -2,3 +2,5 @@ exports.StyleResolverMixin = require('./mixins/style-resolver');
 exports.BrowserStateMixin = require('./mixins/browser-state');
 exports.MatchMediaBase = require('./mixins/match-media-base');
 exports.MatchMediaItem = require('./mixins/match-media-item');
+
+exports.Style = require('./components/style');
